+++ src/main/java/com/corleone/auditoria/repository/LoginHistoricoRepository.java
@@ -3,5 +3,5 @@ package com.corleone.auditoria.repository;
 import com.corleone.auditoria.entity.LoginHistorico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoginHistoricoRepository extends JpaRepository<LoginHistorico, Long> {
+public interface LoginHistoricoRepository extends JpaRepository<LoginHistorico, Integer> {
 }
